# Capstone 5/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Cost Optimizer API
  version: 1.0.0
  description: Automated Cloud Cost Optimizer - 10x Solution Capstone
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /optimizations:
    get:
      summary: List optimization actions
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [PENDING, APPROVED, EXECUTED, REJECTED] }
        - name: riskLevel
          in: query
          schema: { type: string, enum: [LOW, MEDIUM, HIGH] }
        - name: anomalyId
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Paginated optimization actions
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaginatedOptimizations' }
  /optimizations/{id}/approve:
    post:
      summary: Approve optimization action
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approvedBy]
              properties:
                approvedBy: { type: string }
      responses:
        '200':
          description: Approved action
  /optimizations/{id}/reject:
    post:
      summary: Reject optimization action
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Rejected action
...
components:
  schemas:
...
    OptimizationAction:
      type: object
      properties:
        _id: { type: string }
        anomalyId: { type: string }
        actionType: { type: string, enum: [STOP, DOWNSIZE, RIGHTSIZE, DELETE, SCHEDULE] }
        recommendedConfig: { type: object }
        estimatedSavings: { type: number }
        riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH] }
        status: { type: string, enum: [PENDING, APPROVED, EXECUTED, REJECTED] }
        approvedBy: { type: string }
        executedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PaginatedOptimizations:
      type: object
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/OptimizationAction' }
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        pages: { type: integer }